New testrunner supporting more than just ints
* Rewrote the entire ingame testing suite.
* `[MCTest(..)]` can now accept any type, not just ints, and does the type's `==` comparison.
It's very wonky though -- it won't cast automatically, and there's the general issue of numeric types.
- REGRESSION: There's no comparison anymore. For this I need a function that pretty-prints any type.

+++ Compiler/IngameTests/UIntTests.cs
@@ -3,46 +3,46 @@
 namespace MinecraftTests {
     internal class UIntTests {
 
-        [MCTest(26098562)]
-        public static int TestDivision1() {
+        [MCTest(26098562u)]
+        public static uint TestDivision1() {
             uint a = 3210123210u;
             uint b = 123u;
-            return (int)(a / b);
+            return a / b;
         }
 
-        [MCTest(1)]
-        public static int TestDivision2() {
+        [MCTest(1u)]
+        public static uint TestDivision2() {
             uint a = 3210123210u;
             uint b = 2345678910u;
-            return (int)(a / b);
+            return a / b;
         }
 
-        [MCTest(0)]
-        public static int TestDivision3() {
+        [MCTest(0u)]
+        public static uint TestDivision3() {
             uint a = 123u;
             uint b = 3210123210u;
-            return (int)(a / b);
+            return a / b;
         }
 
-        [MCTest(84)]
-        public static int TestMod1() {
+        [MCTest(84u)]
+        public static uint TestMod1() {
             uint a = 3210123210u;
             uint b = 123u;
-            return (int)(a % b);
+            return a % b;
         }
 
-        [MCTest(864444300)]
-        public static int TestMod2() {
+        [MCTest(864444300u)]
+        public static uint TestMod2() {
             uint a = 3210123210u;
             uint b = 2345678910u;
-            return (int)(a % b);
+            return a % b;
         }
 
-        [MCTest(123)]
-        public static int TestMod3() {
+        [MCTest(123u)]
+        public static uint TestMod3() {
             uint a = 123u;
             uint b = 3210123210u;
-            return (int)(a % b);
+            return a % b;
         }
     }
 }
